2006-06-04  Michael Dominic K.  <michaldominik@gmail.com>
	* sample/DecodeBinTranscoder.cs: Fixing ambigious cast.

	* gstreamer-sharp/glue/clock.c:
	* gstreamer-sharp/Clock.custom: Changing the gint64 to guint64 and
	long to ulong respectively. The clocktime (GstClockTime) is measured
	in nanoseconds as unsigned long inside GStreamer. 
	This is introducing a small schizophrenia since some functions (such
	as queries, seeks) take signed long (-1 to mark infinity or error).
	This is the expected Gst behaviour.
	Just sticking silently to long (instead of ulong) will bork us 	sooner
	or later, especially for custom-build low level routines/elements.
	Comments welcome.



git-svn-id: svn://anonsvn.mono-project.com/source/branches/abock/gstreamer-sharp@61438 e3ebcda4-bce8-0310-ba0a-eca2169e7518

+++ sample/DecodeBinTranscoder.cs
@@ -184,8 +184,8 @@ public class DecodeBinTranscoder : IDisposable
         
         transcoder.Progress += delegate(object o, ProgressArgs args) {
             Console.Write("\rEncoding: {0} / {1} ({2:00.00}%) ", 
-                new TimeSpan((args.Position / Clock.GstSecond) * TimeSpan.TicksPerSecond), 
-                new TimeSpan((args.Duration / Clock.GstSecond) * TimeSpan.TicksPerSecond),
+                new TimeSpan((args.Position / (long) Clock.GstSecond) * TimeSpan.TicksPerSecond), 
+                new TimeSpan((args.Duration / (long) Clock.GstSecond) * TimeSpan.TicksPerSecond),
                 ((double)args.Position / (double)args.Duration) * 100.0);
         };
